requirements.txt
arrow==1.3.0
asgiref==3.7.2
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
binaryornot==0.4.4
cachetools==5.3.3
certifi==2024.2.2
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cookiecutter==2.6.0
cryptography==44.0.0
daphne==4.1.2
decouple==0.0.7
Django==5.0.2
django-browser-reload==1.12.1
django-cors-headers==4.3.1
django-decouple==2.1
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
emoji==2.12.1
hyperlink==21.0.0
idna==3.6
incremental==24.7.2
inflection==0.5.1
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgpack==1.1.0
mysqlclient==2.2.4
packaging==24.0
pillow==10.2.0
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
redis==5.2.1
requests==2.31.0
rich==13.7.1
rsa==4.9
sentry-sdk==1.40.6
service-identity==24.2.0
setuptools==75.6.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
Twisted==24.11.0
txaio==23.1.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
zope.interface==7.2
